Move publication schema validation logic to service
