+++ packages/publications/src/Models/PublicationType.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace Hyde\Publications\Models;
 
-use Hyde\Facades\Filesystem;
-use Illuminate\Contracts\Validation\Validator;
 use function array_filter;
 use function array_merge;
 use function dirname;
@@ -24,7 +22,6 @@ use function json_decode;
 use function json_encode;
 use RuntimeException;
 use function str_starts_with;
-use function validator;
 
 /**
  * @see \Hyde\Publications\Testing\Feature\PublicationTypeTest
@@ -228,69 +225,6 @@ class PublicationType implements SerializableContract
      */
     public function validateSchemaFile(bool $throw = true): array
     {
-        $schema = json_decode(Filesystem::getContents($this->getSchemaFile()));
-        $errors = [];
-
-        $schemaValidator = validator([
-            'name' => $schema->name ?? null,
-            'canonicalField' => $schema->canonicalField ?? null,
-            'detailTemplate' => $schema->detailTemplate ?? null,
-            'listTemplate' => $schema->listTemplate ?? null,
-            'sortField' => $schema->sortField ?? null,
-            'sortAscending' => $schema->sortAscending ?? null,
-            'pageSize' => $schema->pageSize ?? null,
-            'fields' => $schema->fields ?? null,
-            'directory' => $schema->directory ?? null,
-        ], [
-            'name' => 'required|string',
-            'canonicalField' => 'nullable|string',
-            'detailTemplate' => 'nullable|string',
-            'listTemplate' => 'nullable|string',
-            'sortField' => 'nullable|string',
-            'sortAscending' => 'nullable|boolean',
-            'pageSize' => 'nullable|integer',
-            'fields' => 'nullable|array',
-            'directory' => 'nullable|prohibited',
-        ]);
-
-        $errors['schema'] = $schemaValidator->errors()->toArray();
-
-        if ($throw) {
-            $schemaValidator->validate();
-        }
-
-        // TODO warn if fields are empty?
-
-        // TODO warn if canonicalField does not match meta field or actual?
-
-        // TODO Warn if template files do not exist (assuming files not vendor views)?
-
-        // TODO warn if pageSize is less than 0 (as that equals no pagination)?
-
-        $errors['fields'] = [];
-
-        foreach ($schema->fields as $field) {
-            $fieldValidator = validator([
-                'type' => $field->type ?? null,
-                'name' => $field->name ?? null,
-                'rules' => $field->rules ?? null,
-                'tagGroup' => $field->tagGroup ?? null,
-            ], [
-                'type' => 'required|string',
-                'name' => 'required|string',
-                'rules' => 'nullable|array',
-                'tagGroup' => 'nullable|string',
-            ]);
-
-            // TODO check tag group exists?
-
-            $errors['fields'][] = $fieldValidator->errors()->toArray();
-
-            if ($throw) {
-                $fieldValidator->validate();
-            }
-        }
-
-        return $errors;
+        return PublicationService::validateSchemaFile($this->getSchemaFile(), $throw);
     }
 }
