+++ packages/publications/src/PublicationService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Hyde\Publications;
 
+use Hyde\Facades\Filesystem;
 use function glob;
 use Hyde\Hyde;
 use Hyde\Publications\Models\PublicationPage;
@@ -11,6 +12,8 @@ use Hyde\Publications\Models\PublicationTags;
 use Hyde\Publications\Models\PublicationType;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
+use function json_decode;
+use function validator;
 
 /**
  * @see \Hyde\Publications\Testing\Feature\PublicationServiceTest
@@ -87,6 +90,79 @@ class PublicationService
     public static function publicationTypeExists(string $pubTypeName): bool
     {
         return static::getPublicationTypes()->has(Str::slug($pubTypeName));
+    }
+
+    /**
+     * Validate the schema.json file is valid.
+     *
+     * @internal This method is experimental and may be removed without notice
+     */
+    public static function validateSchemaFile(string $schemaFile, bool $throw = true): array
+    {
+        $schema = json_decode(Filesystem::getContents($schemaFile));
+        $errors = [];
+
+        $schemaValidator = validator([
+            'name' => $schema->name ?? null,
+            'canonicalField' => $schema->canonicalField ?? null,
+            'detailTemplate' => $schema->detailTemplate ?? null,
+            'listTemplate' => $schema->listTemplate ?? null,
+            'sortField' => $schema->sortField ?? null,
+            'sortAscending' => $schema->sortAscending ?? null,
+            'pageSize' => $schema->pageSize ?? null,
+            'fields' => $schema->fields ?? null,
+            'directory' => $schema->directory ?? null,
+        ], [
+            'name' => 'required|string',
+            'canonicalField' => 'nullable|string',
+            'detailTemplate' => 'nullable|string',
+            'listTemplate' => 'nullable|string',
+            'sortField' => 'nullable|string',
+            'sortAscending' => 'nullable|boolean',
+            'pageSize' => 'nullable|integer',
+            'fields' => 'nullable|array',
+            'directory' => 'nullable|prohibited',
+        ]);
+
+        $errors['schema'] = $schemaValidator->errors()->toArray();
+
+        if ($throw) {
+            $schemaValidator->validate();
+        }
+
+        // TODO warn if fields are empty?
+
+        // TODO warn if canonicalField does not match meta field or actual?
+
+        // TODO Warn if template files do not exist (assuming files not vendor views)?
+
+        // TODO warn if pageSize is less than 0 (as that equals no pagination)?
+
+        $errors['fields'] = [];
+
+        foreach ($schema->fields as $field) {
+            $fieldValidator = validator([
+                'type' => $field->type ?? null,
+                'name' => $field->name ?? null,
+                'rules' => $field->rules ?? null,
+                'tagGroup' => $field->tagGroup ?? null,
+            ], [
+                'type' => 'required|string',
+                'name' => 'required|string',
+                'rules' => 'nullable|array',
+                'tagGroup' => 'nullable|string',
+            ]);
+
+            // TODO check tag group exists?
+
+            $errors['fields'][] = $fieldValidator->errors()->toArray();
+
+            if ($throw) {
+                $fieldValidator->validate();
+            }
+        }
+
+        return $errors;
     }
 
     protected static function getSchemaFiles(): array
